Refactor Quiz Setting Screen: Move components, add expandable text field, handle permissions, and improve UI.
This commit refactors the Quiz Setting Screen for better organization and functionality:

- **Component Extraction:** Moves `QuizSettingTopAppBar`, `ReminderSettingsSection`, and `TagsSettingsSection` into a new `QuizSettingComponents.kt` file for better modularity.
- **Expandable Text Field:** Introduces `ExpandableOutlinedTextField` for the quiz description, allowing users to expand/collapse the text field.
- **Notification Permissions:** Improves notification permission handling in `ReminderSettingsSection` with rationale dialog and snackbar for denied permissions.
- **UI Enhancements:** Adds "See more/less" for the description field and moves SnackbarHostState to the call site for better management.
- **Back Handling:** Retains back press handling with exit confirmation for unsaved changes.
