+++ app/src/main/res/values-vi/strings.xml
@@ -341,4 +341,11 @@
     <string name="reminder_channel_description">Thông báo nhắc bạn ôn tập các quiz</string>
     <string name="reminder_notification_text">Đến lúc ôn tập: %1$s</string> <!-- %1$s will be the quiz title -->
 
+    <string name="see_more">Xem thêm</string>
+    <string name="see_less">Thu gọn</string>
+    <string name="permission_rationale_title">Cần quyền thông báo</string>
+    <string name="permission_rationale_message">Để đặt lời nhắc cho bài kiểm tra này, ứng dụng cần quyền gửi thông báo. Vui lòng cấp quyền này.</string>
+    <string name="permission_rationale_confirm">Cấp quyền</string>
+    <string name="permission_denied_snackbar">Quyền thông báo đã bị từ chối. Không thể đặt lời nhắc.</string>
+
 </resources>
